Fix compiler complains because of uint32_t change
The toolchain used for ESP-IDF 5.x typedefs uint32_t as unsigned long. As a result the printf() format strings using %u (or %d) don't match anymore. The official solution is to use %"PRIu32" but this makes the format strings pretty unreadable. Therefore this change rather casts the printf() parameters to unsigned (or int).

See https://docs.espressif.com/projects/esp-idf/en/v5.0.1/esp32/migration-guides/release-5.x/gcc.html#espressif-toolchain-changes

+++ lib/libssh/src/auth.c
@@ -1573,11 +1573,11 @@ SSH_PACKET_CALLBACK(ssh_packet_userauth_info_request) {
     }
 
     SSH_LOG(SSH_LOG_DEBUG,
-            "%d keyboard-interactive prompts", nprompts);
+            "%d keyboard-interactive prompts", (int) nprompts);
     if (nprompts > KBDINT_MAX_PROMPT) {
         ssh_set_error(session, SSH_FATAL,
                 "Too much prompts requested by the server: %u (0x%.4x)",
-                nprompts, nprompts);
+                (unsigned) nprompts, (unsigned) nprompts);
         ssh_kbdint_free(session->kbdint);
         session->kbdint = NULL;
